+++ views/about.pug
@@ -5,12 +5,12 @@ block content
   background.masthead.text-center.text-white.d-flex
     .container.my-auto
       .row
-      form(method="post" action="saveLink")
-        .form-group
-          h2.section-heading.text-left.text-white.label(for='exampleInputEmail1') User Code
-          input#exampleInputEmail1.form-control(name='url' type='url' aria-describedby='urlHelp' placeholder='Enter Code')
-      form(method="post" action="saveStudent")
+        p1 #{response}
+      form(method="post" action="sendUserData")
         .form-group
           h2.section-heading.text-left.text-white.label(for='exampleInputEmail1') Ngrok https
-          input#exampleInputEmail1.form-control(name='student' placeholder='Enter Ngrok https')
-        a.btn.btn-light.btn-xl.js-scroll-trigger(href='#services') Get Started!
+          input.form-control(name='ngrok' placeholder='Enter Ngrok http')
+          h2.section-heading.text-left.text-white.label(for='exampleInputEmail1') User Code
+          input.form-control(name='code' placeholder='Enter Code')
+        //a.btn.btn-light.btn-xl.js-scroll-trigger(type='submit') Get Started!
+        button.btn.btn-light(type='submit') Submit
