+++ views/layoutBS.pug
@@ -19,12 +19,12 @@ html(lang="en")
               a.nav-link.js-scroll-trigger(href="/") Home
             li.nav-item
               a.nav-link.js-scroll-trigger(href="/about") About
-            li.nav-item
-              a.nav-link.js-scroll-trigger(href="/skills") Projects
-            li.nav-item
-              a.nav-link.js-scroll-trigger(href="/skills") Templates
-            li.nav-item
-              a.nav-link.js-scroll-trigger(href="/login") Login
+            //- li.nav-item
+            //-   a.nav-link.js-scroll-trigger(href="/skills") Projects
+            //- li.nav-item
+            //-   a.nav-link.js-scroll-trigger(href="/skills") Templates
+            //- li.nav-item
+            //-   a.nav-link.js-scroll-trigger(href="/login") Login
 
     block content
     script(src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous")
